GPIO Driver Development Complete

+++ stm32f4xx_drivers/drivers/Inc/stm32f407xx_gpio_driver.h
@@ -120,7 +120,7 @@ void GPIO_ToggleOutputPin(GPIO_RegDef_t *pGPIOx, uint8_t PinNumber);            
  * ISR Configuration and Interrupt Handling
  */
 void GPIO_IRQInterruptConfig(uint8_t IRQNumber, uint8_t EnorDi);            // To configure Interrupt request number
-void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint8_t IRQPriority);
+void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint32_t IRQPriority);
 void GPIO_IRQHandling(uint8_t PinNumber);                                              // To Handle the Interrupt
